1st-ai-test/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
jinja2>=3.1.4
python-dotenv>=1.0.1
requests>=2.32.3
kagglehub>=0.2.9
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.26.4
scikit-learn>=1.5.0
Pillow>=10.3.0
opencv-python>=4.10.0.84
matplotlib>=3.9.0
